Extract identifier_t to separate header file
